Refactor CallDurationCounter environments initialization
